intinty timer

+++ src/main/java/jhaturanga/model/timer/TimerFactory.java
@@ -9,10 +9,10 @@ public interface TimerFactory {
 
     Timer defaultTimer(List<Player> players);
 
-    Timer equalTimer(List<Player> players, int duration);
+    Timer equalTimer(List<Player> players, double duration);
 
-    Timer incrementableTimer(List<Player> players, int duration, int increment);
+    Timer incrementableTimer(List<Player> players, double duration, double increment);
 
-    Timer fromTimerMap(Map<Player, Integer> durations);
+    Timer fromTimerMap(Map<Player, Double> durations);
 
 }
